AArch64: Port Dalvik to 64bit
Change Dalvik such that it compiles and runs as a 64bit program.
Supports only the C interpreter and builds only on the 64bit host.
The VM runs ordinary Dalvik bytecode. The interpreter has been changed to have
64bit registers, but the Java types remain unchanged. The garbage collector,
JNI and interpreter are functional.

+++ vm/BitVector.cpp
@@ -104,7 +104,7 @@ void dvmSetBit(BitVector* pBits, unsigned int num)
     if (num >= pBits->storageSize * sizeof(u4) * 8) {
         if (!pBits->expandable) {
             ALOGE("Attempt to set bit outside valid range (%d, limit is %d)",
-                num, pBits->storageSize * sizeof(u4) * 8);
+                num, (int) (pBits->storageSize * sizeof(u4) * 8));
             dvmAbort();
         }
 
@@ -113,7 +113,7 @@ void dvmSetBit(BitVector* pBits, unsigned int num)
         assert(newSize > pBits->storageSize);
         pBits->storage = (u4*)realloc(pBits->storage, newSize * sizeof(u4));
         if (pBits->storage == NULL) {
-            ALOGE("BitVector expansion to %d failed", newSize * sizeof(u4));
+            ALOGE("BitVector expansion to %d failed", (int) (newSize * sizeof(u4)) );
             dvmAbort();
         }
         memset(&pBits->storage[pBits->storageSize], 0x00,
